@@ -20,7 +20,7 @@ extraReqs = {
     'memcached': ['pylibmc>=1.5.1 ; platform_system != "Windows"'],
     'redis': ['redis>=4.5.5'],
     'converter': [f'large-image-converter{limit_version}'],
-    'colormaps': ['matplotlib', 'tol_colors'],
+    'colormaps': ['matplotlib', 'tol-colors>=2.0.0'],
     'jupyter': ['aiohttp', 'ipyvue', 'ipyleaflet'],
     'tiledoutput': ['pyvips'],
     'performance': [
@@ -78,13 +78,13 @@ setup(
     classifiers=[
         'Development Status :: 5 - Production/Stable',
         'Programming Language :: Python :: 3',
-        'Programming Language :: Python :: 3.9',
         'Programming Language :: Python :: 3.10',
         'Programming Language :: Python :: 3.11',
         'Programming Language :: Python :: 3.12',
         'Programming Language :: Python :: 3.13',
+        'Programming Language :: Python :: 3.14',
     ],
-    python_requires='>=3.9',
+    python_requires='>=3.10',
     install_requires=[
         'cachetools',
         'palettable',
