@@ -14,6 +14,7 @@
 #  limitations under the License.
 #############################################################################
 
+import contextlib
 import hashlib
 import io
 import math
@@ -69,10 +70,8 @@ def _adjustParams(params):
 
     :param params: the request parameters.  May be modified.
     """
-    try:
+    with contextlib.suppress(Exception):
         userAgent = cherrypy.request.headers.get('User-Agent', '').lower()
-    except Exception:
-        pass
     if params.get('encoding', 'JPEG') == 'JPEG':
         if ('ipad' in userAgent or 'ipod' in userAgent or 'iphone' in userAgent or
                 re.match('((?!chrome|android).)*safari', userAgent, re.IGNORECASE)):
@@ -951,6 +950,12 @@ class TilesItemResource(ItemResource):
         .param('frame', 'For multiframe images, the 0-based frame number.  '
                'This is ignored on non-multiframe images.', required=False,
                dataType='int')
+        .param('frameList', 'Comma-separated list of frame numbers to query. '
+               'If provided, returns a list of pixel values for each frame.',
+               required=False)
+        .notes('When frameList is provided, returns a list of pixel '
+               'dictionaries, one for each frame. Otherwise returns a '
+               'single pixel dictionary.')
         .errorResponse('ID was invalid.')
         .errorResponse('Read access was denied for the item.', 403),
     )
@@ -964,7 +969,34 @@ class TilesItemResource(ItemResource):
             ('bottom', float, 'region', 'bottom'),
             ('units', str, 'region', 'units'),
             ('frame', int),
+            ('frameList', str),
         ])
+
+        # Check if frameList is provided for batch frame query
+        if 'frameList' in params:
+            frameList = [
+                int(f.strip()) for f in str(params['frameList']).lstrip(
+                    '[').rstrip(']').split(',')]
+            # Remove frameList and frame from params
+            params.pop('frameList', None)
+            params.pop('frame', None)
+
+            # Get pixel for each frame
+            results = []
+            for frameNum in frameList:
+                frameParams = params.copy()
+                frameParams['frame'] = frameNum
+                try:
+                    pixel = self.imageItemModel.getPixel(item, **frameParams)
+                    pixel['frame'] = frameNum  # Include frame number in result
+                    results.append(pixel)
+                except TileGeneralError as e:
+                    raise RestException(e.args[0])
+                except ValueError as e:
+                    raise RestException('Value Error: %s' % e.args[0])
+            return results
+
+        # Otherwise, use single-frame behavior
         try:
             pixel = self.imageItemModel.getPixel(item, **params)
         except TileGeneralError as e:
