@@ -100,6 +100,14 @@ class AnnotationSchema:
                    r'rgba\(\d+,\s*\d+,\s*\d+,\s*(\d?\.|)\d+\))$',
     }
 
+    patternSchema = {
+        'type': 'string',
+        'description':
+            'If one of "circle", "triangle", "diamond", "flower(number)", '
+            '"star(number)", "jack(number)" where number is from 1 to 16, '
+            'fill closed polylines with a pattern in their strokeColor.',
+    }
+
     colorRangeSchema = {
         'type': 'array',
         'items': colorSchema,
@@ -208,6 +216,7 @@ class AnnotationSchema:
                 'minimum': 0,
             },
             'fillColor': colorSchema,
+            'pattern': patternSchema,
         },
         'required': ['type', 'center', 'radius'],
         'additionalProperties': False,
@@ -243,6 +252,7 @@ class AnnotationSchema:
                     'minItems': 3,
                 },
             },
+            'pattern': patternSchema,
         },
         'required': ['type', 'points'],
         'additionalProperties': False,
@@ -266,6 +276,7 @@ class AnnotationSchema:
             },
             'normal': coordSchema,
             'fillColor': colorSchema,
+            'pattern': patternSchema,
         },
         'decription': 'normal is the positive z-axis unless otherwise '
                       'specified',
@@ -656,6 +667,10 @@ class Annotation(AccessControlledModel):
             ], {}),
             '_version',
             'updated',
+            ([
+                ('_active', SortDir.ASCENDING),
+                ('_elementCount', SortDir.ASCENDING),
+            ], {}),
         ])
         self.ensureTextIndex({
             'annotation.name': 10,
@@ -663,7 +678,7 @@ class Annotation(AccessControlledModel):
         })
 
         self.exposeFields(AccessType.READ, (
-            'annotation', '_version', '_elementQuery', '_active',
+            'annotation', '_version', '_elementQuery', '_active', '_elementCount', '_detailsCount',
         ) + self.baseFields)
         events.bind('model.item.remove', 'large_image_annotation', self._onItemRemove)
         events.bind('model.item.copy.prepare', 'large_image_annotation', self._prepareCopyItem)
@@ -718,6 +733,7 @@ class Annotation(AccessControlledModel):
         destItemId = destItem['_id']
         folder = Folder().load(destItem['folderId'], force=True)
         count = 0
+        user = None
         for annotation in annotations:
             logger.info('Copying annotation %d of %d from %s to %s',
                         count + 1, total, srcItemId, destItemId)
@@ -733,7 +749,11 @@ class Annotation(AccessControlledModel):
             # as the item's folder.
             annotation.pop('access', None)
             self.copyAccessPolicies(destItem, annotation, save=False)
-            self.setPublic(annotation, folder.get('public'), save=False)
+            self.setPublic(annotation, folder.get('public') or False, save=False)
+            if user is None:
+                user = User().load(folder['creatorId'], force=True)
+            if user is not None:
+                self.setUserAccess(annotation, user, AccessType.ADMIN, force=True, save=False)
             self.save(annotation)
             count += 1
         logger.info('Copied %d annotations from %s to %s ',
@@ -751,6 +771,47 @@ class Annotation(AccessControlledModel):
         # Check that all annotations have groups
         for annotation in self.collection.find({'groups': {'$exists': False}}):
             self.injectAnnotationGroupSet(annotation)
+        threading.Thread(target=self._migrateDatabaseBackground, daemon=True).start()
+
+    def _migrateDatabaseBackground(self):
+        needed = self.find({
+            '_active': {'$ne': False},
+            '$or': [
+                {'_elementCount': {'$exists': False}},
+                {'_detailsCount': {'$exists': False}},
+            ],
+        })
+        count = needed.count()
+        if count:
+            logger.info(f'Adding {count} count/details record(s) to existing annotations')
+            lastlog = 0
+            for idx, annot in enumerate(needed):
+                try:
+                    entry = next(Annotationelement().collection.aggregate([{
+                        '$match': {'_version': annot['_version']},
+                    }, {
+                        '$group': {
+                            '_id': None,
+                            '_elementCount': {'$sum': 1},
+                            '_detailsCount': {'$sum': {'$ifNull': ['$bbox.details', 1]}},
+                        },
+                    }]))
+                except StopIteration:
+                    entry = {'_elementCount': 0, '_detailsCount': 0}
+                if time.time() - lastlog > 10:
+                    logger.info(
+                        f'Adding {idx}/{count} count/detail record for '
+                        f'{str(annot["_id"])}, version {annot["_version"]}, '
+                        f'count {entry["_elementCount"]}, details '
+                        f'{entry["_detailsCount"]}')
+                    lastlog = time.time()
+                self.collection.update_one(
+                    {'_id': annot['_id']},
+                    {'$set': {
+                        '_elementCount': entry['_elementCount'],
+                        '_detailsCount': entry['_detailsCount'],
+                    }},
+                )
 
     def _migrateACL(self, annotation):
         """
@@ -940,12 +1001,13 @@ class Annotation(AccessControlledModel):
         _elementQuery = annotation.pop('_elementQuery', None)
         annotation.pop('_active', None)
         annotation.pop('_annotationId', None)
+        annotation.pop('_versionId', None)
         if annotation['annotation'] and not annotation['annotation'].get('name'):
             now = datetime.datetime.now(datetime.timezone.utc)
             annotation['annotation']['name'] = now.strftime('Annotation %Y-%m-%d %H:%M')
 
         def replaceElements(query, doc, *args, **kwargs):
-            Annotationelement().updateElements(doc)
+            elementCount, detailsCount = Annotationelement().updateElements(doc)
             elements = doc['annotation'].pop('elements', None)
             if self._historyEnabled:
                 oldAnnotation = self.collection.find_one(query)
@@ -953,6 +1015,9 @@ class Annotation(AccessControlledModel):
                     oldAnnotation['_annotationId'] = oldAnnotation.pop('_id')
                     oldAnnotation['_active'] = False
                     insert_one(oldAnnotation)
+            if elements is not None:
+                doc['_elementCount'] = elementCount
+                doc['_detailsCount'] = detailsCount
             ret = replace_one(query, doc, *args, **kwargs)
             if elements:
                 doc['annotation']['elements'] = elements
@@ -965,10 +1030,13 @@ class Annotation(AccessControlledModel):
             # the annotation without elements, then restore the elements.
             doc.setdefault('_id', ObjectId())
             if doc['annotation'].get('elements') is not None:
-                Annotationelement().updateElements(doc)
+                elementCount, detailsCount = Annotationelement().updateElements(doc)
             # If we are inserting, we shouldn't have any old elements, so don't
             # bother removing them.
             elements = doc['annotation'].pop('elements', None)
+            if elements is not None:
+                doc['_elementCount'] = elementCount
+                doc['_detailsCount'] = detailsCount
             ret = insert_one(doc, *args, **kwargs)
             if elements is not None:
                 doc['annotation']['elements'] = elements
