@@ -22,6 +22,8 @@ import pickle
 import threading
 import time
 
+import bson.codec_options
+import bson.raw_bson
 import pymongo
 from girder_large_image.models.image_item import ImageItem
 
@@ -81,6 +83,20 @@ class Annotationelement(Model):
             ([
                 ('annotationId', SortDir.ASCENDING),
                 ('_version', SortDir.DESCENDING),
+                ('bbox.size', SortDir.DESCENDING),
+            ], {}),
+            ([
+                ('annotationId', SortDir.ASCENDING),
+                ('_version', SortDir.DESCENDING),
+                ('bbox.lowx', SortDir.ASCENDING),
+                ('bbox.highx', SortDir.ASCENDING),
+                ('bbox.size', SortDir.DESCENDING),
+            ], {}),
+            ([
+                ('annotationId', SortDir.ASCENDING),
+                ('_version', SortDir.DESCENDING),
+                ('bbox.lowy', SortDir.ASCENDING),
+                ('bbox.highy', SortDir.ASCENDING),
                 ('bbox.size', SortDir.DESCENDING),
             ], {}),
         ])
@@ -223,12 +239,14 @@ class Annotationelement(Model):
             'annotationId': annotation.get('_annotationId', annotation['_id']),
             '_version': annotation['_version'],
         }
+        includeCount = True
         for key in region:
             if key in self.bboxKeys and self.bboxKeys[key][1]:
                 if self.bboxKeys[key][1] == '$gte' and float(region[key]) <= 0:
                     continue
                 query[self.bboxKeys[key][0]] = {
                     self.bboxKeys[key][1]: float(region[key])}
+                includeCount = False
         if region.get('sort') in self.bboxKeys:
             sortkey = self.bboxKeys[region['sort']][0]
         else:
@@ -240,17 +258,24 @@ class Annotationelement(Model):
         queryLimit = max(minElements, maxDetails) if maxDetails and (
             not limit or max(minElements, maxDetails) < limit) else limit
         offset = int(region['offset']) if region.get('offset') else 0
-        logger.debug('element query %r for %r', query, region)
-        fields = {'_id': True, 'element': True, 'bbox.details': True, 'datafile': True}
         centroids = str(region.get('centroids')).lower() == 'true'
+        # Specifying a limit helps mongo choose a better index
+        if maxDetails:
+            queryLimit = max(maxDetails, minElements) if queryLimit is None else min(
+                queryLimit, max(maxDetails, minElements))
         if centroids:
-            # fields = {'_id': True, 'element': True, 'bbox': True}
             fields = {
                 '_id': True,
                 'element.id': True,
-                'bbox': True}
+                'bbox.lowx': True,
+                'bbox.lowy': True,
+                'bbox.highx': True,
+                'bbox.highy': True,
+                'bbox.size': True,
+            }
             proplist = []
-            propskeys = ['type', 'fillColor', 'lineColor', 'lineWidth', 'closed']
+            # MUST match below
+            propskeys = ('type', 'fillColor', 'lineColor', 'lineWidth', 'closed')
             # This should match the javascript
             defaultProps = {
                 'fillColor': 'rgba(0,0,0,0)',
@@ -263,22 +288,54 @@ class Annotationelement(Model):
             info['centroids'] = True
             info['props'] = proplist
             info['propskeys'] = propskeys
-        elif region.get('bbox'):
-            fields.pop('bbox.details')
-            fields['bbox'] = True
-        if bbox:
-            fields.pop('bbox.details', None)
-            fields['bbox'] = True
-        elementCursor = self.find(
-            query=query, sort=[(sortkey, sortdir)], limit=queryLimit,
-            offset=offset, fields=fields)
+        else:
+            # Note that it is faster to get all of bbox rather than just
+            # bbox.details (this is not true for centroids)
+            fields = {'_id': True, 'element': True, 'bbox': True, 'datafile': True}
+        logger.debug('element query %r (%r) for %r', query, fields, region)
+        if centroids:
+            elementCursor = self.find(
+                query=query, sort=[(sortkey, sortdir)], limit=queryLimit,
+                offset=offset, fields=fields)
+        else:
+            # By using raw bson to some extent, we save some decoding time
+            # from bson to python.  It isn't clear to me why this reduces
+            # decoding time by 25% or more, but it seems consistent.  When
+            # applied to the centoids, this was actually much slower.
 
+            class SemiRawDocument(bson.raw_bson.RawBSONDocument):
+                def _decode(self, val):
+                    if isinstance(val, bson.raw_bson.RawBSONDocument):
+                        return self._decode(bson.decode(val.raw))
+                    if isinstance(val, dict):
+                        return {k: self._decode(v) for k, v in val.items()}
+                    if isinstance(val, list):
+                        if not val:
+                            return val
+                        if isinstance(val[0], (bson.raw_bson.RawBSONDocument, dict)):
+                            return [self._decode(v) for v in val]
+                    return val
+
+                def __getitem__(self, key):
+                    if key in {'element', 'bbox'}:
+                        if hasattr(self, key):
+                            return getattr(self, key)
+                        val = self._decode(super().__getitem__(key))
+                        setattr(self, key, val)
+                        return val
+                    return super().__getitem__(key)
+
+            elementCursor = self.collection.with_options(
+                codec_options=bson.codec_options.CodecOptions(document_class=SemiRawDocument)).find(
+                    filter=query, sort=[(sortkey, sortdir)], limit=queryLimit,
+                    skip=offset, projection=fields)
         info.update({
-            'count': elementCursor.count(),
             'offset': offset,
             'filter': query,
             'sort': [sortkey, sortdir],
         })
+        if includeCount:
+            info['count'] = elementCursor.count()
         details = count = 0
         if maxDetails:
             info['maxDetails'] = maxDetails
@@ -288,19 +345,23 @@ class Annotationelement(Model):
             info['limit'] = limit
         for entry in elementCursor:
             element = entry['element']
-            element.setdefault('id', entry['_id'])
             if centroids:
                 bbox = entry.get('bbox')
                 if not bbox or 'lowx' not in bbox or 'size' not in bbox:
                     continue
-                prop = tuple(
-                    element.get(key, defaultProps.get(key)) for key in propskeys
-                    if element.get(key, defaultProps.get(key)) is not None)
+                # MUST match above; faster than
+                #   prop = tuple(element.get(key) for key in propskeys)
+                prop = (
+                    element.get('type'),
+                    element.get('fillColor'),
+                    element.get('lineColor'),
+                    element.get('lineWidth'),
+                    element.get('closed'))
                 if prop not in props:
                     props[prop] = len(props)
-                    proplist.append(list(prop))
+                    proplist.append([element.get(key, defaultProps.get(key)) for key in propskeys])
                 yield [
-                    str(element['id']),
+                    element.get('id', entry['_id']),
                     (bbox['lowx'] + bbox['highx']) / 2,
                     (bbox['lowy'] + bbox['highy']) / 2,
                     bbox['size'] if entry.get('type') != 'point' else 0,
@@ -308,6 +369,7 @@ class Annotationelement(Model):
                 ]
                 details += 1
             else:
+                element.setdefault('id', entry['_id'])
                 if entry.get('datafile'):
                     datafile = entry['datafile']
                     chunksize = 1024 ** 2
@@ -618,6 +680,8 @@ class Annotationelement(Model):
                     len(entries), time.time() - chunkStartTime, prepTime,
                     chunk + len(entries), len(elements)))
                 lastTime = time.time()
+            return len(entries), sum(
+                el.get('bbox', {}).get('details') or 1 for el in entries)
         except Exception:
             logger.exception('Failed to update element chunk')
             raise
@@ -632,18 +696,26 @@ class Annotationelement(Model):
         startTime = time.time()
         elements = annotation['annotation'].get('elements', [])
         if not len(elements):
-            return
+            return 0, 0
         now = datetime.datetime.now(datetime.timezone.utc)
         threads = large_image.config.cpu_count()
         chunkSize = int(max(100000 // threads, 10000))
         insertLock = threading.Lock()
+        count, details = 0, 0
         with concurrent.futures.ThreadPoolExecutor(max_workers=threads) as pool:
+            futures = []
             for chunk in range(0, len(elements), chunkSize):
-                pool.submit(self.updateElementChunk, elements, chunk,
-                            chunkSize, annotation, now, insertLock)
+                futures.append(pool.submit(
+                    self.updateElementChunk, elements, chunk, chunkSize,
+                    annotation, now, insertLock))
+            for future in concurrent.futures.as_completed(futures):
+                chunkCount, chunkDetails = future.result()
+                count += chunkCount
+                details += chunkDetails
         if time.time() - startTime > 10:
             logger.info('inserted %d elements in %4.2fs' % (
                 len(elements), time.time() - startTime))
+        return count, details
 
     def getElementGroupSet(self, annotation):
         query = {
