@@ -24,6 +24,8 @@
 
 import atexit
 import builtins
+import contextlib
+import importlib.metadata
 import logging
 import math
 import os
@@ -33,8 +35,6 @@ import threading
 import types
 import weakref
 import zipfile
-from importlib.metadata import PackageNotFoundError
-from importlib.metadata import version as _importlib_version
 
 import numpy as np
 
@@ -45,11 +45,8 @@ from large_image.constants import TILE_FORMAT_NUMPY, SourcePriority
 from large_image.exceptions import TileSourceError, TileSourceFileNotFoundError
 from large_image.tilesource import FileTileSource, nearPowerOfTwo
 
-try:
-    __version__ = _importlib_version(__name__)
-except PackageNotFoundError:
-    # package is not installed
-    pass
+with contextlib.suppress(importlib.metadata.PackageNotFoundError):
+    __version__ = importlib.metadata.version(__name__)
 
 bioformats = None
 # import javabridge
@@ -78,14 +75,10 @@ def _monitor_thread():
             while len(_openImages):
                 source = _openImages.pop()
                 source = source()
-                try:
+                with contextlib.suppress(Exception):
                     source._bioimage.close()
-                except Exception:
-                    pass
-                try:
+                with contextlib.suppress(Exception):
                     source._bioimage = None
-                except Exception:
-                    pass
         except Exception:
             pass
         finally:
@@ -247,11 +240,9 @@ class BioformatsFileTileSource(FileTileSource, metaclass=LruCacheMetaclass):
                     # release some file handles.
                     self._bioimage.init_reader()
                 except Exception as exc:
-                    try:
+                    with contextlib.suppress(Exception):
                         # Ask to open a file that should never exist
                         self._bioimage.rdr.setId('__\0__')
-                    except Exception:
-                        pass
                     self._bioimage.close()
                     self._bioimage = None
                     raise exc
@@ -336,9 +327,10 @@ class BioformatsFileTileSource(FileTileSource, metaclass=LruCacheMetaclass):
         try:
             self._lastGetTileException = 'raise'
             self.getTile(0, 0, self.levels - 1)
-            delattr(self, '_lastGetTileException')
+            del self._lastGetTileException
         except Exception as exc:
             raise TileSourceError('Bioformats cannot read a tile: %r' % exc)
+        self._checkForOffset()
         self._populatedLevels = len([
             v for v in self._metadata['frameSeries'][0]['series'] if v is not None])
 
@@ -356,6 +348,40 @@ class BioformatsFileTileSource(FileTileSource, metaclass=LruCacheMetaclass):
                 with _javabridgeAttachLock:
                     if javabridge.get_env():
                         javabridge.detach()
+
+    def _checkForOffset(self):
+        """
+        The bioformats DICOM reader does unfortunate things to MONOCHROME1
+        16-bit images.  Store an offset to undo it, if appropriate.
+        """
+        if self._metadata.get('readerClassName') != 'loci.formats.in.DicomReader':
+            return
+        if self._metadata.get('seriesMetadata', {}).get(
+                '0028,0004 Photometric Interpretation') != 'MONOCHROME1':
+            return
+        if np.issubdtype(self.dtype, np.uint8):
+            self._fix_offset = 255
+            return
+        if not np.issubdtype(self.dtype, np.int16) and not np.issubdtype(self.dtype, '>i2'):
+            return
+        # This is bioformats behavior
+        try:
+            maxPixelRange = int(self._metadata['seriesMetadata'].get(
+                '0028,1051 Window Width', 0))
+        except Exception:
+            maxPixelRange = -1
+        try:
+            centerPixelValue = int(self._metadata['seriesMetadata'].get(
+                '0028,1050 Window Center', 0))
+        except Exception:
+            centerPixelValue = -1
+        maxPixelValue = maxPixelRange + (centerPixelValue // 2)
+        maxAllowRange = 2 ** int(self._metadata['seriesMetadata'].get(
+            '0028,0101 Bits Stored', 16)) - 1
+        if maxPixelRange == -1 or centerPixelValue < maxPixelRange // 2:
+            maxPixelValue = maxAllowRange
+        if maxPixelValue:
+            self._fix_offset = maxPixelValue
 
     def _metadataForCurrentSeries(self, rdr):
         self._metadata = getattr(self, '_metadata', {})
@@ -553,11 +579,9 @@ class BioformatsFileTileSource(FileTileSource, metaclass=LruCacheMetaclass):
                 if metadata.get(key):
                     found = re.match(r'^[^0-9.]*(\d*\.?\d+)[^0-9.]+(\d*\.?\d+)\D*$', metadata[key])
                     if found:
-                        try:
+                        with contextlib.suppress(Exception):
                             self._magnification['mm_x'], self._magnification['mm_y'] = (
                                 float(found.groups()[0]) * units, float(found.groups()[1]) * units)
-                        except Exception:
-                            pass
         for key in magkeys:
             if metadata.get(key):
                 self._magnification['magnification'] = float(metadata[key])
@@ -723,6 +747,8 @@ class BioformatsFileTileSource(FileTileSource, metaclass=LruCacheMetaclass):
             retile[0:min(tile.shape[0], finalHeight), 0:min(tile.shape[1], finalWidth)] = tile[
                 0:min(tile.shape[0], finalHeight), 0:min(tile.shape[1], finalWidth)]
             tile = retile
+        if hasattr(self, '_fix_offset') and format == TILE_FORMAT_NUMPY:
+            tile = self._fix_offset - tile
         return self._outputTile(tile, format, x, y, z, pilImageAllowed, numpyAllowed, **kwargs)
 
     def getAssociatedImagesList(self):
