@@ -14,12 +14,12 @@
 #  limitations under the License.
 ##############################################################################
 
+import contextlib
+import importlib.metadata
 import math
 import os
 import threading
 import warnings
-from importlib.metadata import PackageNotFoundError
-from importlib.metadata import version as _importlib_version
 
 import numpy as np
 
@@ -30,11 +30,8 @@ from large_image.tilesource import FileTileSource
 
 nd2 = None
 
-try:
-    __version__ = _importlib_version(__name__)
-except PackageNotFoundError:
-    # package is not installed
-    pass
+with contextlib.suppress(importlib.metadata.PackageNotFoundError):
+    __version__ = importlib.metadata.version(__name__)
 
 
 def _lazyImport():
@@ -321,9 +318,15 @@ class ND2FileTileSource(FileTileSource, metaclass=LruCacheMetaclass):
             fc //= self._nd2sizes[axis]
             tileframe = tileframe[fp // fc]
             fp = fp % fc
-        with self._tileLock:
-            # Have dask use single-threaded since we are using a lock anyway.
-            tile = tileframe[y0:y1:step, x0:x1:step].compute(scheduler='single-threaded').copy()
+        # We had done
+        #   with self._tileLock:
+        #       # Have dask use single-threaded since we are using a lock anyway.
+        #       tile = tileframe[y0:y1:step, x0:x1:step].compute(scheduler='single-threaded').copy()
+        # but this is no longer necessary.  We could limit the number of thread
+        # workers used by dask, but this doesn't actually change much.  We
+        # could also use scheduler='processes', but this fails to close handles
+        # a certain way and prints a lot of warnings and is slower.
+        tile = tileframe[y0:y1:step, x0:x1:step].compute().copy()
         return self._outputTile(tile, TILE_FORMAT_NUMPY, x, y, z,
                                 pilImageAllowed, numpyAllowed, **kwargs)
 
