@@ -14,15 +14,14 @@
 #  limitations under the License.
 #############################################################################
 
+import contextlib
+import importlib.metadata
 import math
 import os
 import pathlib
 import tempfile
 import threading
 import warnings
-from contextlib import suppress
-from importlib.metadata import PackageNotFoundError
-from importlib.metadata import version as _importlib_version
 
 import numpy as np
 import PIL.Image
@@ -36,17 +35,15 @@ from large_image.constants import (PROJECTION_SENTINEL, TILE_FORMAT_IMAGE,
                                    TileOutputMimeTypes)
 from large_image.exceptions import (TileSourceError,
                                     TileSourceFileNotFoundError,
-                                    TileSourceInefficientError)
+                                    TileSourceInefficientError,
+                                    TileSourceXYZRangeError)
 from large_image.tilesource.geo import (GDALBaseFileTileSource,
                                         ProjUnitsAcrossLevel0,
                                         ProjUnitsAcrossLevel0_MaxSize)
 from large_image.tilesource.utilities import JSONDict
 
-try:
-    __version__ = _importlib_version(__name__)
-except PackageNotFoundError:
-    # package is not installed
-    pass
+with contextlib.suppress(importlib.metadata.PackageNotFoundError):
+    __version__ = importlib.metadata.version(__name__)
 
 rio = None
 Affine = None
@@ -93,7 +90,7 @@ class RasterioFileTileSource(GDALBaseFileTileSource, metaclass=LruCacheMetaclass
     cacheName = 'tilesource'
     name = 'rasterio'
 
-    def __init__(self, path, projection=PROJECTION_SENTINEL, unitsPerPixel=None, **kwargs):
+    def __init__(self, path, projection=PROJECTION_SENTINEL, unitsPerPixel=None, **kwargs):  # noqa
         """Initialize the tile class.
 
         See the base class for other available parameters.
@@ -178,7 +175,32 @@ class RasterioFileTileSource(GDALBaseFileTileSource, metaclass=LruCacheMetaclass
             logY = math.log(float(self.sizeY) / self.tileHeight)
             computedLevel = math.ceil(max(logX, logY) / math.log(2))
             self.sourceLevels = self.levels = int(max(0, computedLevel) + 1)
-
+            with self._getDatasetLock:
+                try:
+                    subdatasets = self.dataset.subdatasets
+                except (RuntimeError, ZeroDivisionError):
+                    subdatasets = None
+                if subdatasets and len(subdatasets) > 1:
+                    self._frames = []
+                    for sdsrecord in subdatasets:
+                        sdspath = os.path.join(os.path.dirname(self._largeImagePath), sdsrecord)
+                        try:
+                            sds = rio.open(sdspath)
+                        except Exception:
+                            sdspath = sdsrecord
+                            try:
+                                sds = rio.open(sdspath)
+                            except Exception:
+                                sds = None
+                        try:
+                            if (sds and sds.width == self.sourceSizeX and
+                                    sds.height == self.sourceSizeY and
+                                    self.dataset.indexes == sds.indexes):
+                                self._frames.append(sdspath)
+                        except Exception:
+                            pass
+                    if len(self._frames) <= 1:
+                        del self._frames
             self._unitsPerPixel = unitsPerPixel
             self.projection is None or self._initWithProjection(unitsPerPixel)
             self._getPopulatedLevels()
@@ -206,19 +228,19 @@ class RasterioFileTileSource(GDALBaseFileTileSource, metaclass=LruCacheMetaclass
         :param analysisSize: optional default to 1024
         :param onlyMinMax: optional default to True
         """
-        # default frame to 0 in case it is set to None from outside
-        frame = frame or 0
+        frame = self._getFrame(frame=frame or 0)
+        ds = self._getFrameDataset(frame)
 
         # read band information
-        bandInfo = self.getBandInformation()
+        bandInfo = self.getBandInformation(dataset=ds)
 
         # get the minmax value from the band
         hasMin = all(b.get('min') is not None for b in bandInfo.values())
         hasMax = all(b.get('max') is not None for b in bandInfo.values())
-        if not frame and onlyMinMax and hasMax and hasMin:
+        if onlyMinMax and hasMax and hasMin:
             with self._getDatasetLock:
                 dtype = self.dataset.profile['dtype']
-            self._bandRanges[0] = {
+            self._bandRanges[frame] = {
                 'min': np.array([b['min'] for b in bandInfo.values()], dtype=dtype),
                 'max': np.array([b['max'] for b in bandInfo.values()], dtype=dtype),
             }
@@ -478,7 +500,7 @@ class RasterioFileTileSource(GDALBaseFileTileSource, metaclass=LruCacheMetaclass
             return self._bandInfo
 
         # check if the dataset is cached
-        cache = not dataset
+        cache = not dataset or dataset == self.dataset
 
         # do everything inside the dataset lock to avoid multiple read
         with self._getDatasetLock:
@@ -551,6 +573,9 @@ class RasterioFileTileSource(GDALBaseFileTileSource, metaclass=LruCacheMetaclass
                 'sourceBounds': self.getBounds(),
                 'bands': self.getBandInformation(),
             })
+        if hasattr(self, '_frames'):
+            metadata['frames'] = [{} for _ in self._frames]
+            self._addMetadataFrameInformation(metadata)
         return metadata
 
     def getInternalMetadata(self, **kwargs):
@@ -591,8 +616,30 @@ class RasterioFileTileSource(GDALBaseFileTileSource, metaclass=LruCacheMetaclass
 
         return result
 
+    def _getFrameDataset(self, frame):
+        ds = self.dataset
+        if frame or hasattr(self, '_frames'):
+            if frame < 0 or not hasattr(self, '_frames') or frame >= len(self._frames):
+                msg = 'Frame does not exist'
+                raise TileSourceXYZRangeError(msg)
+            if not hasattr(self, '_frameCache') or not hasattr(self, '_frameCacheMaxSize'):
+                self._frameCache = {}
+                self._frameCacheMaxSize = 10
+            if frame in self._frameCache:
+                ds = self._frameCache[frame]
+            else:
+                with self._getDatasetLock:
+                    if len(self._frameCache) >= self._frameCacheMaxSize:
+                        self._frameCache = {}
+                    ds = rio.open(self._frames[frame])
+                    self._frameCache[frame] = ds
+        return ds
+
     @methodcache()
     def getTile(self, x, y, z, pilImageAllowed=False, numpyAllowed=False, **kwargs):
+        frame = self._getFrame(**kwargs)
+        ds = self._getFrameDataset(frame)
+        tile = None
         if not self.projection:
             self._xyzInRange(x, y, z)
             factor = int(2 ** (self.levels - 1 - z))
@@ -605,13 +652,20 @@ class RasterioFileTileSource(GDALBaseFileTileSource, metaclass=LruCacheMetaclass
 
             with self._getDatasetLock:
                 window = rio.windows.Window(xmin, ymin, xmax - xmin, ymax - ymin)
-                count = self.dataset.count
-                tile = self.dataset.read(
-                    window=window,
-                    out_shape=(count, h, w),
-                    resampling=rio.enums.Resampling.nearest,
-                )
-
+                count = ds.count
+                try:
+                    tile = ds.read(
+                        window=window,
+                        out_shape=(count, h, w),
+                        resampling=rio.enums.Resampling.nearest,
+                    )
+                except Exception:
+                    tile = None
+                    self.logger.exception('Failed to getTile')
+                    if hasattr(self, '_frameCache'):
+                        self._frameCache = {}
+                    else:
+                        self.dataset = rio.open(self._largeImagePath)
         else:
             xmin, ymin, xmax, ymax = self.getTileCorners(z, x, y)
             bounds = self.getBounds(self.projection)
@@ -635,28 +689,29 @@ class RasterioFileTileSource(GDALBaseFileTileSource, metaclass=LruCacheMetaclass
             # Adding an alpha band when the source has one is trouble.
             # It will result in surprisingly unmasked data.
             src_alpha_band = 0
-            for i, interp in enumerate(self.dataset.colorinterp):
+            for i, interp in enumerate(ds.colorinterp):
                 if interp == rio.enums.ColorInterp.alpha:
                     src_alpha_band = i
             add_alpha = not src_alpha_band
 
             # read the image as a warp vrt
-            with self._getDatasetLock:
-                with rio.vrt.WarpedVRT(
-                    self.dataset,
-                    resampling=rio.enums.Resampling.nearest,
-                    crs=self.projection,
-                    transform=dst_transform,
-                    height=self.tileHeight,
-                    width=self.tileWidth,
-                    add_alpha=add_alpha,
-                ) as vrt:
-                    try:
-                        tile = vrt.read(resampling=rio.enums.Resampling.nearest)
-                    except Exception:
-                        self.logger.exception('Failed to getTile')
-                        tile = np.zeros((1, 1))
+            with self._getDatasetLock, rio.vrt.WarpedVRT(
+                ds,
+                resampling=rio.enums.Resampling.nearest,
+                crs=self.projection,
+                transform=dst_transform,
+                height=self.tileHeight,
+                width=self.tileWidth,
+                add_alpha=add_alpha,
+            ) as vrt:
+                try:
+                    tile = vrt.read(resampling=rio.enums.Resampling.nearest)
+                except Exception:
+                    self.logger.exception('Failed to getTile')
+                    tile = None
 
+        if tile is None:
+            tile = np.zeros((1, 1))
         # necessary for multispectral images:
         # set the coordinates first and the bands at the end
         if len(tile.shape) == 3:
@@ -788,7 +843,7 @@ class RasterioFileTileSource(GDALBaseFileTileSource, metaclass=LruCacheMetaclass
 
         # check if the units is a string or projection material
         isProj = False
-        with suppress(rio.errors.CRSError):
+        with contextlib.suppress(rio.errors.CRSError):
             isProj = make_crs(units) is not None
 
         # convert the coordinates if a projection exist
@@ -925,11 +980,13 @@ class RasterioFileTileSource(GDALBaseFileTileSource, metaclass=LruCacheMetaclass
                 x, y = self.toNativePixelCoordinates(x, y)
 
             if 0 <= int(x) < self.sizeX and 0 <= int(y) < self.sizeY:
+                frame = self._getFrame(**kwargs)
+                ds = self._getFrameDataset(frame)
                 with self._getDatasetLock:
-                    for i in self.dataset.indexes:
+                    for i in ds.indexes:
                         window = rio.windows.Window(int(x), int(y), 1, 1)
                         try:
-                            value = self.dataset.read(
+                            value = ds.read(
                                 i, window=window, resampling=rio.enums.Resampling.nearest,
                             )
                             value = value[0][0]  # there should be 1 single pixel
@@ -958,6 +1015,8 @@ class RasterioFileTileSource(GDALBaseFileTileSource, metaclass=LruCacheMetaclass
         :returns: regionData, formatOrRegionMime: the image data and either the
             mime type, if the format is TILE_FORMAT_IMAGE, or the format.
         """
+        frame = self._getFrame(**kwargs)
+        ds = self._getFrameDataset(frame)
         # cast format as a tuple if needed
         format = format if isinstance(format, (tuple, set, list)) else (format,)
 
@@ -1002,7 +1061,7 @@ class RasterioFileTileSource(GDALBaseFileTileSource, metaclass=LruCacheMetaclass
             dst_transform = Affine(xres, 0.0, left, 0.0, -yres, top)
 
             with rio.vrt.WarpedVRT(
-                self.dataset,
+                ds,
                 resampling=rio.enums.Resampling.nearest,
                 crs=self.projection,
                 transform=dst_transform,
@@ -1011,7 +1070,7 @@ class RasterioFileTileSource(GDALBaseFileTileSource, metaclass=LruCacheMetaclass
             ) as vrt:
                 data = vrt.read(resampling=rio.enums.Resampling.nearest)
 
-            profile = self.dataset.meta.copy()
+            profile = ds.meta.copy()
             profile.update(
                 large_image.tilesource.utilities._rasterioParameters(
                     defaultCompression='lzw', **kwargs,
@@ -1027,8 +1086,8 @@ class RasterioFileTileSource(GDALBaseFileTileSource, metaclass=LruCacheMetaclass
                 dst.write(data)
                 # Write colormaps if available
                 for i in range(data.shape[0]):
-                    if self.dataset.colorinterp[i].name.lower() == 'palette':
-                        dst.write_colormap(i + 1, self.dataset.colormap(i + 1))
+                    if ds.colorinterp[i].name.lower() == 'palette':
+                        dst.write_colormap(i + 1, ds.colormap(i + 1))
 
             return outputPath, TileOutputMimeTypes['TILED']
 
