@@ -1,6 +1,8 @@
 import concurrent.futures
+import contextlib
 import datetime
 import fractions
+import importlib.metadata
 import json
 import logging
 import math
@@ -9,8 +11,6 @@ import re
 import struct
 import threading
 import time
-from importlib.metadata import PackageNotFoundError
-from importlib.metadata import version as _importlib_version
 from tempfile import TemporaryDirectory
 
 import numpy as np
@@ -25,11 +25,8 @@ from . import format_aperio
 
 pyvips = None
 
-try:
-    __version__ = _importlib_version(__name__)
-except PackageNotFoundError:
-    # package is not installed
-    pass
+with contextlib.suppress(importlib.metadata.PackageNotFoundError):
+    __version__ = importlib.metadata.version(__name__)
 
 
 logger = logging.getLogger('large-image-converter')
@@ -589,7 +586,7 @@ def _convert_large_image_frame(
         preferredVipsCast=preferredVipsCast, **kwargs)
 
 
-def _output_type(lidata):  # noqa
+def _output_type(lidata, keepFloat=False):  # noqa
     """
     Determine how to cast and scale vips data based on actual image contents.
     """
@@ -599,6 +596,10 @@ def _output_type(lidata):  # noqa
         return None
     if intype == np.uint8 or intype == np.uint16:
         return None
+    if keepFloat and np.issubdtype(intype, np.floating):
+        if intype == np.float16 or intype == np.float32:
+            return (pyvips.BandFormat.FLOAT, 0, 1)
+        return (pyvips.BandFormat.DOUBLE, 0, 1)
     logger.debug('Checking data range')
     minval = maxval = None
     for frame in range(len(lidata['metadata'].get('frames', [0]))):
@@ -657,7 +658,7 @@ def _convert_large_image(inputPath, outputPath, tempPath, lidata, **kwargs):
         images.
     """
     ts = lidata['tilesource']
-    lidata['_vips_cast'] = _output_type(lidata)
+    lidata['_vips_cast'] = _output_type(lidata, kwargs.get('keepFloat', False))
     numFrames = len(lidata['metadata'].get('frames', [0]))
     outputList = []
     tasks = []
@@ -986,6 +987,8 @@ def convert(inputPath, outputPath=None, **kwargs):  # noqa: C901
         primary ifds.
     :param overwrite: if not True, throw an exception if the output path
         already exists.
+    :param keepFloat: if True, keep float or double data types as they are, if
+        possible.
 
     Additional optional parameters:
 
